requirements.txt
glfw==2.8.0
numpy==2.2.4
pillow==11.1.0
PyOpenGL==3.1.9
